+++ .gitignore
@@ -1,1 +1,2 @@
-.idea
+.idea,
+/node_modules
