+++ index.html
@@ -43,6 +43,6 @@
             </div>
         </div>
         
-        <script src="app.js"></script>
+        <script src="dist/main.js"></script>
     </body>
 </html>
